Introduce basic LINC-Switch integration
LINC-Switch is integrated with mininet by introducing a new class that extends
Switch. For the time being LINC-Switch support is not fully tested a need
to be extended so that all mininet features can work with LINC. To start
mininet with LINC simply run the mininet executable as follows:
`sudo mn --switch=linc`.
(Motivated by FlowForwarding/LINC-Switch#129).